remove createAmplifyAdapter & refactor
